3D: combine plasma solver and beam calculator for cpu and gpu
