--- lcode2dPy/beam3d_gpu/__init__.py
@@ -1,4 +1,0 @@
-"""Beam storage and interaction module for LCODE 3D."""
-from .beam_calculator import BeamCalculator
-from .beam_io import BeamSource, BeamDrain
-from .data import BeamParticles, concatenate_beam_layers
