+++ lcode2dPy/plasma3d/solver.py
@@ -1,15 +1,16 @@
 """Module for calculator of one full step along xi."""
 from ..config.config import Config
-from .data import Fields, Currents, Particles, Const_Arrays
-from .fields import FieldComputer
-from .move import ParticleMover
-from .rhoj import RhoJComputer
+from ..plasma3d.rhoj import get_rhoj_computer
+from ..plasma3d.fields import get_field_computer
+from ..plasma3d.move import get_plasma_particles_mover
+
 
 class Plane2d3vPlasmaSolver(object):
     def __init__(self, config: Config):
-        self.FComputer = FieldComputer(config)
-        self.PMover = ParticleMover(config)
-        self.CComputer = RhoJComputer(config)
+        self.compute_rhoj = get_rhoj_computer(config)
+        self.compute_fields = get_field_computer(config)
+        self.move_particles, self.move_particles_wo_fields = \
+            get_plasma_particles_mover(config)
 
     # Perfoms one full step along xi.
     # To understand the numerical scheme, read values as following:
@@ -18,42 +19,41 @@ class Plane2d3vPlasmaSolver(object):
     # *_full = * on the next xi step (fullstep), an index number = k + 1
     # *_prevprev = * on the xi step with an index number k - 1
     def step_dxi(
-        self, particles_prev: Particles, fields_prev: Fields,
-        currents_prev: Currents, const_arrays: Const_Arrays,
+        self, particles_prev, fields_prev, currents_prev, const_arrays,
         rho_beam_full, rho_beam_prev
     ):
-        particles_full = self.PMover.move_particles_wo_fields(particles_prev)
+        particles_full = self.move_particles_wo_fields(particles_prev)
 
 
-        particles_full = self.PMover.move_particles(
+        particles_full = self.move_particles(
             fields_prev, particles_prev, particles_full
         )
-        currents_full = self.CComputer.compute_rhoj(
+        currents_full = self.compute_rhoj(
             particles_full, const_arrays
         )
 
-        _, fields_half = self.FComputer.compute_fields(
+        _, fields_half = self.compute_fields(
             fields_prev, fields_prev, const_arrays, rho_beam_full, rho_beam_prev,
             currents_prev, currents_full
         )
 
 
-        particles_full = self.PMover.move_particles(
+        particles_full = self.move_particles(
             fields_half, particles_prev, particles_full
         )
-        currents_full = self.CComputer.compute_rhoj(
+        currents_full = self.compute_rhoj(
             particles_full, const_arrays
         )
 
-        fields_full, fields_half = self.FComputer.compute_fields(
+        fields_full, fields_half = self.compute_fields(
             fields_half, fields_prev, const_arrays, rho_beam_full, rho_beam_prev,
             currents_prev, currents_full
         )
 
-        particles_full = self.PMover.move_particles(
+        particles_full = self.move_particles(
             fields_half, particles_prev, particles_full
         )
-        currents_full = self.CComputer.compute_rhoj(
+        currents_full = self.compute_rhoj(
             particles_full, const_arrays
         )
